* Feature: Large maps with improved map generation
  * Map generation now supports sizes up to 1000x1000 tiles.
    Very large maps can take a long time to generate and may reduce performance.
  * Maps that differ from the original 80x50 size require the new savegame format.
    Older savegames are not compatible with non-80x50 map sizes, because the legacy format assumes fixed map dimensions.
  * The "Customize World" screen now includes larger size presets up to 160x100 (Huge):
    * Tiny (40x25), Small (60x40), Normal (80x50), Large (120x75), Huge (160x100)
    * You can also enter a custom size between 20x20 and 1000x1000.
    * The standard "New Game" menu still starts a normal 80x50 map for quick play.
  * If generation is still running and the intro is skipped, the intro screen shows a progress message.
  * When generation is complete, the intro text turns green.
  * If generation fails (for example due to an exception), the intro screen shows an error message and then retries map generation automatically.
    Logs still contain the technical details for troubleshooting.
  * Generation logs now include stage timings, making it easier to understand runtime and spot bottlenecks.
  * Number of continents is only limited by `int` range (2 147 483 647); the previous cap of 15 (from the original 4-bit nibble format) is removed.

+++ src/ContinentTraversalDelegate.cs
@@ -18,16 +18,16 @@ namespace CivOne
         private readonly int _height;
         private readonly int[,] _relativePositions;
         private readonly Func<int, int, bool> _isOceanAt;
-        private readonly Func<int, int, byte> _getContinentIdAt;
-        private readonly Action<int, int, byte> _setContinentIdAt;
+        private readonly Func<int, int, int> _getContinentIdAt;
+        private readonly Action<int, int, int> _setContinentIdAt;
 
         internal ContinentTraversalDelegate(
             int width,
             int height,
             int[,] relativePositions,
             Func<int, int, bool> isOceanAt,
-            Func<int, int, byte> getContinentIdAt,
-            Action<int, int, byte> setContinentIdAt)
+            Func<int, int, int> getContinentIdAt,
+            Action<int, int, int> setContinentIdAt)
         {
             _width = width;
             _height = height;
@@ -37,7 +37,7 @@ namespace CivOne
             _setContinentIdAt = setContinentIdAt;
         }
 
-        internal ulong CountContinent(int x, int y, bool ocean, byte continentId)
+        internal ulong CountContinent(int x, int y, bool ocean, int continentId)
         {
             if (_height <= 0 || _width <= 0)
             {
